requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
# Django==5.0.1
django-heroku==0.3.1
django-import-export==3.3.5
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
